+++ app/pb_kits/playbook/pb_bar_graph/bar_graph.rb
@@ -19,7 +19,9 @@ module Playbook
                                 default:[]
       prop :y_axis_min, type: Playbook::Props::Numeric
       prop :y_axis_max, type: Playbook::Props::Numeric
-
+      prop :legend, type: Playbook::Props::Boolean,
+                    default: false
+      prop :height
 
       def chart_type
         orientation == "horizontal" ? "bar" : "column"
@@ -37,6 +39,8 @@ module Playbook
           xAxisCategories: x_axis_categories,
           yAxisMin: y_axis_min,
           yAxisMax: y_axis_max,
+          legend: legend,
+          height: height,
         }.to_json.html_safe
       end