+++ app/pb_kits/playbook/pb_bar_graph/docs/index.js
@@ -1,1 +1,3 @@
 export { default as BarGraphDefault } from './_bar_graph_default.jsx'
+export { default as BarGraphLegend } from './_bar_graph_legend.jsx'
+export { default as BarGraphHeight } from './_bar_graph_height.jsx'