+++ app/pb_kits/playbook/pb_star_rating/star_rating.rb
@@ -9,11 +9,12 @@ module Playbook
 
       prop :hide_rating, type: Playbook::Props::Boolean,
                          default: false
+
       prop :rating, type: Playbook::Props::Numeric,
                     default: 0
 
       def star_count
-        [*1..rating.to_f]
+        rating.floor
       end
 
       def star_full