+++ CHANGELOG.md
@@ -7,9 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-### Added
+## [4.16.0] 2020-5-28
 
+### Added
 - Add Title Count React kit ([#825](https://github.com/powerhome/playbook/pull/825) @kellyeryan)
+- Add React Star Rating kit ([#822](https://github.com/powerhome/playbook/pull/822) @kellyeryan)
+- Added legend prop to Bar Graph Kit and Line Graph Kit ([#818](https://github.com/powerhome/playbook/pull/818)@christinaatai)
+- Add React to LabelPill kit ([#813](https://github.com/powerhome/playbook/pull/813) @kellyeryan)
+- Updated Bar Graph Kit and Line Graph Kit with height prop ([#819](https://github.com/powerhome/playbook/pull/819)@christinaatai)
+- Added link & new_window props to circle_icon_button rails/react ([#823](https://github.com/powerhome/playbook/pull/823)@kre8sions)
+- Add Selectable Icon and Selectable Card Icon kits ([#782](https://github.com/powerhome/playbook/pull/782))
+- Create Source React Playbook Kit ([#810](https://github.com/powerhome/playbook/pull/810) @kellyeryan)
+
+### Fixed
+- Updated index.js to include ProgressStepItem ([#831](https://github.com/powerhome/playbook/pull/831)@christinaatai)
+- Fix Rails hide_rating feature so that it hides the rating ([#822](https://github.com/powerhome/playbook/pull/822) @kellyeryan)
+- Added aria, data, id props to Text Input kit on both React and Ruby sides.([#812](https://github.com/powerhome/playbook/pull/812)@kellyeryan)
+- Add aria to Ruby LabelPill kit ([#813](https://github.com/powerhome/playbook/pull/813) @kellyeryan)
+- Fix Currency kit performance issues (Rails) ([#820](https://github.com/powerhome/playbook/pull/820) @thestephenmarshall)
 
 ## [4.15.0] 2020-5-22